#97 Add tracing info log on battle start and trigger hostility check on player reconnect
